add signup page

+++ server/controller/vendor.js
@@ -8,6 +8,7 @@ const Products = require("../models/product.js");
 module.exports.updateProfile = async (req, res) => {
     let { model, field, fieldUpdateValue } = req.query; // model specifies user or vendor, field specifies the field to be updated, fieldUpdateValue specifies the update to be made
 
+    if(model || field || fieldUpdateValue) return res.sendStatus(400);
     // console.log(`model: ${model}, field: ${field}, fieldUpdateValue: ${fieldUpdateValue}`);
 
     // update field with fieldUpdateValue
@@ -20,8 +21,14 @@ module.exports.updateProfile = async (req, res) => {
     }
 
     // keeping cached data up to date
-    const updatedUser = await fetchVendorDetailsFromDB(req.user.user);
-    await setCache(`vendor:${req.user.userId}`, updatedUser);
+    try{
+        const updatedUser = await fetchVendorDetailsFromDB(req.user.user);
+        await setCache(`vendor:${req.user.userId}`, updatedUser);
+    } catch (error) {
+        console.log(error);
+        return res.sendStatus(500);
+    }
+
     return res.sendStatus(200);
 }
 
@@ -30,6 +37,8 @@ module.exports.getProductsOfVendor = async (req, res) => {
     const products = await Products.find({ vendor: req.user._id });
     console.log(products); // array of products
 
+    if(products.length == 0) return res.sendStatus(404);
+
     return res.send(products); //products belonging to vendor
 }
 
@@ -37,8 +46,8 @@ module.exports.addNewProduct = async (req, res) => {
     // add product belonging to vendor
     console.log(req.body);
     const newProduct = new Products({ ...req.body, vendor: req.user._id });
-    await newProduct.save();
-    console.log(newProduct);
+    if (!(await newProduct.save())) return res.sendStatus(500);
+    // console.log(newProduct);
     res.send("httpStatus");
 }
 
@@ -48,24 +57,29 @@ module.exports.newProductFormPage = (req, res) => {
 
 module.exports.getProductDetails = async (req, res) => {
     const { productId } = req.params;
+    if(!productId) return res.sendStatus(400);
     const productDetails = await Products.findById(productId);
+    if(!productDetails) return res.sendStatus(404);
     console.log(productDetails);
-    res.send(productDetails);
+    return res.send(productDetails);
 }
 
 module.exports.updateProductDetails = async (req, res) => {
     const { field, fieldUpdate } = req.query;
     const { productId } = req.params;
+    if(!field || !fieldUpdate || !productId) return res.sendStatus(400);
     console.log(`field: ${field}; fieldUpdate: ${fieldUpdate}; productId: ${productId}`);
-    await Products.findByIdAndUpdate(productId, { $set: { [field]: fieldUpdate } });
+    const update = await Products.findByIdAndUpdate(productId, { $set: { [field]: fieldUpdate } }, { new: true });
     // update product details
     // use query string to specify what detail of a product to update
-    res.send("httpStatus");
+    if(!update) return res.sendStatus(500);
+    res.sendStatus(200);
 }
 
 module.exports.deleteProduct = async (req, res) => {
     // delete product by productId
     const { productId } = req.params;
+    if(!productId) return res.sendStatus(400);
     await Products.findByIdAndDelete(productId);
     res.send("httpStatus");
 }
